refactor: moved common settings forms to respective components
