+++ components/index.ts
@@ -7,3 +7,9 @@ export * from "./ClientsTable";
 export * from "./ClientsForm";
 export * from "./NotesCard";
 export * from "./CollapseNavLinks";
+export * from "./UserLocationForm";
+export * from "./UserDetailsForm";
+export * from "./ThemeSelectionCard";
+export * from "./PaymentForm";
+export * from "./UserBillingForm";
+export * from "./PasswordResetForm";
